Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -20,7 +20,7 @@ int main() {
     }
     else if(c == '/'){
         if(b != 0){
-            devide = float(a/b);
+            devide = a/b;
             printf("%d\n", devide);
         }
         else{
